LPS-56047 Fix redirect and forward, removing duplication
