+++ portal-impl/src/com/liferay/portlet/messageboards/action/BaseMessageBoardsMVCActionCommand.java
@@ -1,0 +1,53 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portlet.messageboards.action;
+
+import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.util.WebKeys;
+
+import javax.portlet.ActionRequest;
+import javax.portlet.ActionResponse;
+import java.io.IOException;
+
+/**
+ * @author Adolfo Pérez
+ */
+public abstract class BaseMessageBoardsMVCActionCommand
+	extends BaseMVCActionCommand {
+
+	protected void sendRedirect(
+		ActionRequest actionRequest, ActionResponse actionResponse)
+		throws IOException {
+
+		String redirect = getRedirect(actionRequest);
+
+		if (Validator.isNotNull(redirect)) {
+			sendRedirect(actionRequest, actionResponse, redirect);
+		}
+	}
+
+	protected String getRedirect(ActionRequest actionRequest) {
+		String redirect = (String)actionRequest.getAttribute(WebKeys.REDIRECT);
+
+		if (Validator.isNotNull(redirect)) {
+			return redirect;
+		}
+
+		return ParamUtil.getString(actionRequest, "redirect");
+	}
+
+}
